requirements.txt
altair==5.0.1
appdirs==1.4.4
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.1
cycler==0.11.0
decorator==5.1.1
filelock==3.12.2
fonttools==4.40.0
frozendict==2.3.8
fsspec==2023.6.0
fundamentalanalysis==0.2.14
gitdb==4.0.10
GitPython==3.1.31
html5lib==1.1
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.6.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
lxml==4.9.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.1
mdurl==0.1.2
mpmath==1.3.0
multitasking==0.0.11
networkx==3.1
nltk==3.8.1
numpy==1.24.3
packaging==23.1
pandas==2.0.2
Pillow==9.5.0
plotly==5.15.0
protobuf==4.23.2
psycopg2==2.9.7
pyarrow==12.0.1
pycparser==2.21
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.1.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
streamlit==1.23.1
sympy==1.12
tenacity==8.2.2
threadpoolctl==3.1.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
transformers==4.30.2
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3
urllib3==2.0.3
validators==0.20.0
valinvest==0.0.2
watchdog==3.0.0
webencodings==0.5.1
Wikipedia-API==0.5.8
yfinance==0.2.20
zipp==3.15.0
